pyproject.toml
[tool.poetry]
name = "clinicahire"
version = "0.1.0"
description = ""
authors = ["MaximPyanin <maximpyanin@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.15.2"
fastapi = "^0.114.2"
ruff = "^0.6.5"
sqlalchemy = "^2.0.34"
alembic = "^1.13.2"
asyncpg = "^0.29.0"
bcrypt = "^4.2.0"
pyjwt = {extras = ["crypto"], version = "^2.9.0"}
cryptography = "^43.0.1"
httpx = "^0.27.2"
dependency-injector = "^4.42.0"
gunicorn = "^23.0.0"
uvicorn = "^0.30.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
